requirements.txt
django
djangorestframework
request
numpy
terminaltables
pandas
django-rest-swagger
